+++ lib/src/controls/utils/hover_button.dart
@@ -7,6 +7,28 @@ typedef ButtonStateWidgetBuilder = Widget Function(
   Set<ButtonStates> state,
 );
 
+class _HoverButtonInherited extends InheritedWidget {
+  const _HoverButtonInherited({
+    required super.child,
+    required this.states,
+  });
+
+  final Set<ButtonStates> states;
+
+  static _HoverButtonInherited of(BuildContext context) {
+    return context.dependOnInheritedWidgetOfExactType<_HoverButtonInherited>()!;
+  }
+
+  static _HoverButtonInherited? maybeOf(BuildContext context) {
+    return context.dependOnInheritedWidgetOfExactType<_HoverButtonInherited>();
+  }
+
+  @override
+  bool updateShouldNotify(_HoverButtonInherited oldWidget) {
+    return states != oldWidget.states;
+  }
+}
+
 /// Base widget for any widget that requires input.
 class HoverButton extends StatefulWidget {
   /// Creates a hover button.
@@ -27,6 +49,7 @@ class HoverButton extends StatefulWidget {
     this.onHorizontalDragStart,
     this.onHorizontalDragUpdate,
     this.onHorizontalDragEnd,
+    this.gestures = const {},
     this.onFocusTap,
     this.onFocusChange,
     this.autofocus = false,
@@ -59,6 +82,15 @@ class HoverButton extends StatefulWidget {
   final GestureDragStartCallback? onHorizontalDragStart;
   final GestureDragUpdateCallback? onHorizontalDragUpdate;
   final GestureDragEndCallback? onHorizontalDragEnd;
+
+  /// The gestures that this widget will attempt to recognize.
+  ///
+  /// This should be a map from [GestureRecognizer] subclasses to
+  /// [GestureRecognizerFactory] subclasses specialized with the same type.
+  ///
+  /// This value can be late-bound at layout time using
+  /// [RawGestureDetectorState.replaceGestureRecognizers].
+  final Map<Type, GestureRecognizerFactory> gestures;
 
   /// When the button is focused and is actioned, with either the enter or space
   /// keys
@@ -133,8 +165,12 @@ class HoverButton extends StatefulWidget {
   @override
   State<HoverButton> createState() => _HoverButtonState();
 
-  static _HoverButtonState of(BuildContext context) {
-    return context.findAncestorStateOfType<_HoverButtonState>()!;
+  static _HoverButtonInherited of(BuildContext context) {
+    return _HoverButtonInherited.of(context);
+  }
+
+  static _HoverButtonInherited? maybeOf(BuildContext context) {
+    return _HoverButtonInherited.maybeOf(context);
   }
 }
 
@@ -265,6 +301,7 @@ class _HoverButtonState extends State<HoverButton> {
       onHorizontalDragEnd: widget.onHorizontalDragEnd,
       child: widget.builder(context, states),
     );
+    w = RawGestureDetector(gestures: widget.gestures, child: w);
     if (widget.focusEnabled) {
       w = FocusableActionDetector(
         mouseCursor: widget.cursor ?? MouseCursor.defer,
@@ -297,7 +334,6 @@ class _HoverButtonState extends State<HoverButton> {
     w = MergeSemantics(
       child: Semantics(
         label: widget.semanticLabel,
-        button: true,
         enabled: enabled,
         focusable: enabled && node.canRequestFocus,
         focused: node.hasFocus,
@@ -305,6 +341,11 @@ class _HoverButtonState extends State<HoverButton> {
       ),
     );
     if (widget.margin != null) w = Padding(padding: widget.margin!, child: w);
+
+    w = _HoverButtonInherited(
+      states: states,
+      child: w,
+    );
     return w;
   }
 }